Fix empty deleteTimeseries requests (#18012) (#18037)
(cherry picked from commit cd5c3a31195c0ef69b0bf3ccacfd10257bd43ddb)

+++ iotdb-client/session/src/main/java/org/apache/iotdb/session/SessionConnection.java
@@ -779,6 +779,9 @@ public class SessionConnection {
 
   protected void deleteTimeseries(List<String> paths)
       throws IoTDBConnectionException, StatementExecutionException {
+    if (paths.isEmpty()) {
+      return;
+    }
     callWithRetryAndVerify(() -> client.deleteTimeseries(sessionId, paths));
   }
